refactor:브랜치내용 맞추기+컨벤션 적용
