+++ src/main/java/com/plog/plogbackend/domain/Member/MemberService.java
@@ -1,7 +1,7 @@
-package com.plog.plogbackend.domain.member;
+package com.plog.plogbackend.domain.Member;
 
-import com.plog.plogbackend.domain.member.dto.MemberSignupRequest;
-import com.plog.plogbackend.domain.member.dto.MemberSignupResponse;
+import com.plog.plogbackend.domain.Member.dto.MemberSignupRequest;
+import com.plog.plogbackend.domain.Member.dto.MemberSignupResponse;
 import com.plog.plogbackend.security.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
