+++ src/main/java/com/plog/plogbackend/domain/Member/Role.java
@@ -1,4 +1,4 @@
-package com.plog.plogbackend.domain.member;
+package com.plog.plogbackend.domain.Member;
 
 public enum Role {
   ROLE_USER, // 정상 회원가입 완료 회원
