+++ src/main/java/com/plog/plogbackend/domain/Member/Terms.java
@@ -1,4 +1,4 @@
-package com.plog.plogbackend.domain.member;
+package com.plog.plogbackend.domain.Member;
 
 public enum Terms {
   MARKETING // 마케팅 정보 활용 동의;
