+++ src/main/java/com/plog/plogbackend/global/common/Enum/EntityStatus.java
@@ -1,4 +1,4 @@
-package com.plog.plogbackend.global.common.enums;
+package com.plog.plogbackend.global.common.Enum;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
