+++ src/main/java/com/plog/plogbackend/global/common/entity/BaseTimeStatusEntity.java
@@ -1,20 +1,35 @@
 package com.plog.plogbackend.global.common.entity;
 
-import com.plog.plogbackend.global.common.enums.EntityStatus;
+import com.plog.plogbackend.global.common.Enum.EntityStatus;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Getter
 @MappedSuperclass
-public abstract class BaseTimeStatusEntity extends BaseTimeEntity {
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseTimeStatusEntity {
 
+  @CreatedDate
+  @Column(updatable = false, nullable = false)
+  private LocalDateTime createdAt;
+
+  @LastModifiedDate
+  @Column(nullable = false)
+  private LocalDateTime updatedAt;
+
+  // 삭제 시각 기록 (삭제되지 않았을 때는 null)
   private LocalDateTime deletedAt;
 
+  // 공통 상태 관리 (기본값 ACTIVE)
   @Enumerated(EnumType.STRING)
   @Column(nullable = false, length = 20)
-  private EntityStatus status = EntityStatus.ACTIVE; // protected → private
+  protected EntityStatus status = EntityStatus.ACTIVE;
 
+  /** 논리적 삭제 */
   public void deleteEntity() {
     this.status = EntityStatus.DELETED;
     this.deletedAt = LocalDateTime.now();
