+++ src/main/java/com/plog/plogbackend/security/error/OAuth2FailureHandler.java
@@ -2,20 +2,21 @@ package com.plog.plogbackend.security.error;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.stereotype.Component;
 
-import java.io.IOException;
-
 @Slf4j
 @Component
 public class OAuth2FailureHandler implements AuthenticationFailureHandler {
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
-        log.error("소셜 로그인 실패 원인: {}", exception.getMessage(), exception);
+  @Override
+  public void onAuthenticationFailure(
+      HttpServletRequest request, HttpServletResponse response, AuthenticationException exception)
+      throws IOException {
+    log.error("소셜 로그인 실패 원인: {}", exception.getMessage(), exception);
 
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "소셜 로그인 실패");
-    }
+    response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "소셜 로그인 실패");
+  }
 }
